Switched from iframe to jQuery include
Since the iframe wasn't wanting to work on OSX, I switched to using a div (which is easier to fit in with sizing than the iframe anyway) and using jQuery to grab the individual html pages and setting the html into the div I created. I had to change some file paths, but it actually made the modularity easier and more closely resembles an application.